Extracted out ConfSetSubdecks::task to confSetSubdecks
